Added bleed effect
Changelog:
- Added new effect: bleeding. Mobs that bleed take 1 damage point (half-heart) per second.
	- The stronger is the effect, the more damage the mob takes. It capes at 3 damage points, however.
- Added new particle: bleeding. Appears only on players when they have the "bleeding" effect active.
- Added new enchantment: Bloody blade. It can be applied to daggers, swords and axes. Applies Bleeding to the objective for 10 seconds.
	- On level II, applies Bleeding II for 5 seconds.
	- On level III, applies Bleeding III for 5 seconds.
- Daggers have now a "bleed percentage", dependent of the tier. if the chance succeds, applies Bleeding to the objective for 10 seconds.
	However, if the dagger is enchanted with Bleeding, this chance is anulled and applies the enchantment instead.
- Poison touch and Living touch now last 5 seconds, and their max level has been increased to III.
	- The applied effect amplifies 1 level per enchantment level.
- Fine mining and Rough mining can now be applied to axes and shovels. Previusly, only worked on pickaxes.

+++ Enhanched_Minecraft_Mod/src/main/java/net/marmar/enhanced_playthrough/block/alloyfurnace/AbstractAlloyFurnaceBlock.java
@@ -52,7 +52,7 @@ public abstract class AbstractAlloyFurnaceBlock extends BaseEntityBlock implemen
     public void animateTick(BlockState pState, Level pLevel, BlockPos pPos, RandomSource pRandom) {
         if (pState.getValue(BURNING)) {
             double X_position = (double)pPos.getX() + 0.5;
-            double Y_position = (double)pPos.getY();
+            double Y_position = pPos.getY();
             double Z_position = (double)pPos.getZ() + 0.5;
             if (pRandom.nextDouble() < 0.1) {
                 pLevel.playLocalSound(X_position, Y_position, Z_position, SoundEvents.FURNACE_FIRE_CRACKLE, SoundSource.BLOCKS, 1.0F, 1.0F, false);
